CLOUD-1753: Compilation fix
<==== Mercurial data ====>
Branch: cloud_17.2

+++ unit_tests/cloud_db_ut/src/ec2/transaction_log_cache_ut.cpp
@@ -103,7 +103,7 @@ protected:
     {
         ASSERT_EQ(
             *m_transactionSequenceGenerated.begin(),
-            m_cache.committedTimestampSequence());
+            (int)m_cache.committedTimestampSequence());
     }
 
     void assertIfCommittedDataIsNotVisible()
